Fixing output Flux_data for Fit Saved Lines function
